Update the extension tests
