+++ com.unity.ml-agents/Runtime/Unity.ML-Agents.asmdef
@@ -4,7 +4,8 @@
     "references": [
         "Unity.ML-Agents.CommunicatorObjects",
         "Unity.Mathematics",
-        "Unity.InferenceEngine"
+        "Unity.InferenceEngine",
+        "Unity.InputSystem"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
@@ -32,6 +33,11 @@
             "name": "com.unity.modules.physics2d",
             "expression": "1.0.0",
             "define": "MLA_UNITY_PHYSICS2D_MODULE"
+        },
+        {
+            "name": "com.unity.inputsystem",
+            "expression": "1.1.0",
+            "define": "MLA_INPUT_SYSTEM"
         }
     ],
     "noEngineReferences": false
