+++ com.unity.ml-agents/Tests/Runtime/Input/Adaptors/FloatInputActionAdapatorTests.cs
@@ -1,7 +1,6 @@
 #if MLA_INPUT_TESTS
 using NUnit.Framework;
 using Unity.MLAgents.Actuators;
-using Unity.MLAgents.Actuators;
 using UnityEngine;
 using UnityEngine.InputSystem;
