+++ com.unity.ml-agents/Tests/Runtime/Sensor/RigidBodyPoseExtractorTests.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 using NUnit.Framework;
 using Unity.MLAgents.Sensors;
 
-namespace Unity.MLAgents.Tests.Sensors
+namespace Unity.MLAgents.Tests
 {
     public class RigidBodyPoseExtractorTests
     {
