--- com.unity.ml-agents/Tests/Runtime/Sensors.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 88803e617d2d429aa4838bdaaeb64bac
-timeCreated: 1616140923
